# Backend: Comprehensive architecture refactoring - service layer, repository pattern, async operations