--- a/backend/app/repositories/__init__.py
+++ b/backend/app/repositories/__init__.py
@@ -1,0 +1,15 @@
+"""
+Repository pattern implementations for data access layer
+"""
+
+from .base import BaseRepository
+from .repository_repository import RepositoryRepository
+from .technology_repository import TechnologyRepository
+from .research_task_repository import ResearchTaskRepository
+
+__all__ = [
+    "BaseRepository",
+    "RepositoryRepository",
+    "TechnologyRepository",
+    "ResearchTaskRepository",
+]