--- a/backend/app/services/__init__.py
+++ b/backend/app/services/__init__.py
@@ -3,9 +3,17 @@ Service layer for external integrations and business logic
 """
 
 from app.services.github_service import GitHubService
+from app.services.github_async import GitHubAsyncService
 from app.services.rag_service import RAGService
+from app.services.repository_service import RepositoryService
+from app.services.technology_service import TechnologyService
+from app.services.research_service import ResearchService
 
 __all__ = [
     "GitHubService",
+    "GitHubAsyncService",
     "RAGService",
+    "RepositoryService",
+    "TechnologyService",
+    "ResearchService",
 ]